@@ -9,34 +9,33 @@
     "format": "prettier --write ."
   },
   "devDependencies": {
-    "@dicebear/avatars": "^4.5.4",
-    "@dicebear/avatars-gridy-sprites": "^4.5.4",
+    "@dicebear/avatars": "^4.6.3",
+    "@dicebear/avatars-gridy-sprites": "^4.6.3",
     "@sveltejs/adapter-node": "next",
-    "@sveltejs/kit": "^1.0.0-next.84",
+    "@sveltejs/kit": "^1.0.0-next.107",
     "@sveltejs/svelte-virtual-list": "^3.0.1",
-    "@types/classnames": "^2.2.11",
-    "@typescript-eslint/eslint-plugin": "^4.22.0",
-    "@typescript-eslint/parser": "^4.22.0",
+    "@typescript-eslint/eslint-plugin": "^4.23.0",
+    "@typescript-eslint/parser": "^4.23.0",
     "autoprefixer": "^10.2.5",
-    "classnames": "^2.2.6",
-    "cssnano": "^4.1.10",
-    "date-fns": "^2.19.0",
-    "eslint": "^7.24.0",
-    "eslint-config-prettier": "^8.2.0",
-    "eslint-plugin-svelte3": "^3.1.2",
-    "libphonenumber-js": "^1.9.16",
-    "msw": "^0.28.1",
-    "postcss": "^8.2.7",
+    "classnames": "^2.3.1",
+    "cssnano": "^5.0.2",
+    "date-fns": "^2.21.3",
+    "eslint": "^7.26.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-svelte3": "^3.2.0",
+    "libphonenumber-js": "^1.9.17",
+    "msw": "^0.28.2",
+    "postcss": "^8.2.15",
     "postcss-load-config": "^3.0.1",
-    "prettier": "^2.2.1",
-    "prettier-plugin-svelte": "^2.2.0",
-    "svelte": "^3.29.0",
+    "prettier": "^2.3.0",
+    "prettier-plugin-svelte": "^2.3.0",
+    "svelte": "^3.38.2",
     "svelte-i18n": "^3.3.9",
-    "svelte-preprocess": "^4.6.9",
-    "tailwindcss": "^2.1.1",
-    "tslib": "^2.0.0",
-    "typescript": "^4.0.0",
-    "vite": "^2.1.0"
+    "svelte-preprocess": "^4.7.3",
+    "tailwindcss": "^2.1.2",
+    "tslib": "^2.2.0",
+    "typescript": "^4.2.4",
+    "vite": "^2.3.2"
   },
   "type": "module",
   "engines": {
